Fixing chart values

+++ PFM/Commands/DeleteCommand.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Windows;
 using System.Windows.Input;
 
 namespace PFM
@@ -25,12 +26,16 @@ namespace PFM
 
         public void Execute(object parameter)
         {
-            // delete item from DB
-            viewModel.DBInventory.DeleteFromDB(viewModel.DBInventory.SelectedItem);
-            // delete item from list
-            viewModel.DBInventory.InventoryRecords.Remove(viewModel.DBInventory.SelectedItem);
-            // update charts
-            viewModel.UpdateCharts();
+            MessageBoxResult result = MessageBox.Show("Biztosan törli a tételt?", "Tétel törlése", MessageBoxButton.YesNo, MessageBoxImage.Question);
+            if(result == MessageBoxResult.Yes)
+            {
+                // delete item from DB
+                viewModel.DBInventory.DeleteFromDB(viewModel.DBInventory.SelectedItem);
+                // delete item from list
+                viewModel.DBInventory.InventoryRecords.Remove(viewModel.DBInventory.SelectedItem);
+                // update charts
+                viewModel.UpdateCharts();
+            }
         }
     }
 }
